Changed the name of the package from books.model to model.books.
