+++ src/main/java/org/bitmaxsystems/oop2library/model/books/Author.java
@@ -1,25 +1,25 @@
-package org.bitmaxsystems.oop2library.books.model;
+package org.bitmaxsystems.oop2library.model.books;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Table(name = "genres")
+@Table(name = "authors")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class Genre {
+public class Author {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, unique = true)
+    @Column(nullable = false)
     private String name;
 
-    public Genre(String name) {
+    public Author(String name) {
         this.name = name;
     }
 }
