+++ src/main/java/org/bitmaxsystems/oop2library/model/books/BookStatus.java
@@ -1,4 +1,4 @@
-package org.bitmaxsystems.oop2library.books.model;
+package org.bitmaxsystems.oop2library.model.books;
 
 public enum BookStatus {
     AVAILABLE,
